+++ src/functions/public/ConvertTo-SimpleProperty.ps1
@@ -1,0 +1,82 @@
+function ConvertTo-SimpleProperty {
+    #TODO: Only need to input $Value and output the PSCustomObject, drop the other params
+    param (
+        $InputObject,
+
+        [string]$Property,
+
+        [hashtable]$PropertyDictionary = @{},
+
+        [string]$Prefix
+    )
+
+    $Value = $InputObject.$Property
+
+    if ($null -ne $Value) {
+        # We wrap this in an expression and use output redirection to supress this error:
+        # The following exception occurred while retrieving member "GetType": "Not implemented"
+        [string]$Type = & { $Value.GetType().FullName } 2>$null
+    } else {
+        [string]$Type = $null
+    }
+
+    switch ($Type) {
+        'System.DirectoryServices.DirectoryEntry' {
+            $PropertyDictionary["$Prefix$Property"] = ConvertFrom-DirectoryEntry -DirectoryEntry $Value
+        }
+        'System.DirectoryServices.PropertyCollection' {
+            $ThisObject = @{}
+
+            ForEach ($ThisProperty in $Value.Keys) {
+                $ThisPropertyString = ConvertFrom-PropertyValueCollectionToString -PropertyValueCollection $Value[$ThisProperty]
+                $ThisObject[$ThisProperty] = $ThisPropertyString
+
+                # This copies the properties up to the top level.
+                # Want to remove this later
+                # The nested pscustomobject accomplishes the goal of removing hashtables and PropertyValueCollections and PropertyCollections
+                # But I may have existing functionality expecting these properties so I am not yet ready to remove this
+                # When I am, I should move this code into a ConvertFrom-PropertyCollection function in the Adsi module
+                $PropertyDictionary["$Prefix$ThisProperty"] = $ThisPropertyString
+
+            }
+            $PropertyDictionary["$Prefix$Property"] = [PSCustomObject]$ThisObject
+            continue
+        }
+        'System.DirectoryServices.PropertyValueCollection' {
+            $PropertyDictionary["$Prefix$Property"] = ConvertFrom-PropertyValueCollectionToString -PropertyValueCollection $Value
+            continue
+        }
+        'System.Object[]' {
+            $PropertyDictionary["$Prefix$Property"] = $Value
+            continue
+        }
+        'System.Object' {
+            $PropertyDictionary["$Prefix$Property"] = $Value
+            continue
+        }
+        'System.Management.Automation.PSCustomObject' {
+            $PropertyDictionary["$Prefix$Property"] = $Value
+            continue
+        }
+        'System.Collections.Hashtable' {
+            $PropertyDictionary["$Prefix$Property"] = [PSCustomObject]$Value
+            continue
+        }
+        'System.Byte[]' {
+            $PropertyDictionary["$Prefix$Property"] = ConvertTo-DecStringRepresentation -ByteArray $Value
+        }
+        default {
+            <#
+                By default we will just let most types get cast as a string
+                Includes but not limited to:
+                    $null (because GetType is not implemented)
+                    System.String
+                    System.Boolean
+            #>
+            $PropertyDictionary["$Prefix$Property"] = "$Value"
+            continue
+        }
+    }
+
+    return $PropertyDictionary
+}