+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.15] - 2022-07-31 - Bug fixes for Expand-AccountPermission which involved creating ConvertTo-SimpleProperty
+
+## [2.0.14] - 2022-07-31 - Bug fixes for Expand-AccountPermission which involved creating ConvertTo-SimpleProperty
+
+## [2.0.13] - 2022-07-31 - Bug fixes for Expand-AccountPermission which involved creating ConvertTo-SimpleProperty
+
+## [2.0.12] - 2022-07-31 - Bug fixes for Expand-AccountPermission which involved creating ConvertTo-SimpleProperty
+
+## [2.0.11] - 2022-07-31 - Bug fixes for Expand-AccountPermission, this involved creating ConvertTo-SimpleProperty
+
 ## [2.0.10] - 2022-07-31 - Added Find-ServerNameInPath
 
 ## [2.0.9] - 2022-07-30 - Specified -FilterScript for readability